+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Remove redundant `rust-client.nestedMultiRootConfigInOutermost` setting (originally used to work around non-multi-project limitations)
 * Ignore setting `rust-client.enableMultiProjectSetup` (it's always on by default)
 * Fix support for multiple VSCode workspaces